Move connectivity to its own file
